Move XML template to separate file, fix formatting, and use path.
