+++ CMakeLists_files.cmake
@@ -111,6 +111,7 @@ list (APPEND MAIN_SOURCE_FILES
 if (DAMARIS_FOUND)
   list (APPEND MAIN_SOURCE_FILES opm/simulators/utils/DamarisOutputModule.cpp) 
   list (APPEND MAIN_SOURCE_FILES opm/simulators/utils/DamarisKeywords.cpp)
+  list (APPEND MAIN_SOURCE_FILES opm/simulators/utils/initDamarisXMLFile.cpp)
 endif()
 if(CUDA_FOUND)
   list (APPEND MAIN_SOURCE_FILES opm/simulators/linalg/bda/cuda/cusparseSolverBackend.cu)
